Move VariablesExtractorCommand into src/Command dir.
